Working with Actors. Initial outline
- Copied various sections from the Working with actors document to
separate documents (but kept the original document intact)
- Moved all parts concerning ReceiveActor API from the ReceiveActor.md
to a separate document (Defining an actor class).
- Removed the ReceiveActor document because the remaining stuff
detailing the Become API are explained elsewhere as well.
- Creating actors document is missing the F# example. The current F#
document has so many stuff in there.. and since im not an F# dev, im
uncertain whats relevant for the `Creating actors` document, and whats
not. So im leaving that for someone else to fill in.
- Creating actors document contains a segment about the Inbox. Don't
think it belongs there, but im not sure where to put it.
- What's in an Actor document. The outline is there, but im not sure
what to put in there so I put in TODO's
- ActorRefs and ActorSelections document. The outline is there.
- creating-actors-with-props. The outline is there.
- handling-messages. The outline is there. But im unsure what to put in
there. Receive Actor API and UntypedActor API are described elsewhere as
well.
- Sending messages document. Copied the contents from the 'Send
messages' section in the Working with actors document