Removes the smoothscroll class from body
This change allows browsers to decide how they want to handle scrolling rather than forcing it on users. Chrome as of version 49 automatically enables smooth scrolling: https://developers.google.com/web/updates/2016/02/smooth-scrolling-in-chrome-49 with the preference: chrome://flags/#disable-smooth-scrolling and myself and many others disable the feature. Smooth scrolling is like piloting a large boat in the water, where steering is sluggish and there's lag to everything you do. With smooth scrolling turned off, it's like driving a Formula-1 race car: scrolling is snappy and stops exactly when you want it to.